requirements.txt
asgiref==3.5.2
Django==4.1.4
django-admin==2.0.1
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.4
djangorestframework==3.14.0
pytz==2022.6
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
